+++ src/wenmode/plugins/__init__.py
@@ -1,9 +1,17 @@
 """Built-in Wenmode plugins and plugin API types."""
 
-from .types import Plugin as Plugin
-from .types import PluginSpec as PluginSpec
-from .types import PluginTarget as PluginTarget
+from .._declarative import DeclarativePluginSpec as DeclarativePluginSpec
+from .._declarative import InlineDelimited as InlineDelimited
+from .._declarative import RenderTemplate as RenderTemplate
+from .._declarative import install_declarative as install_declarative
 from .types import RendererHandlers as RendererHandlers
 from .types import plugin as plugin
 
-__all__ = ['Plugin', 'PluginSpec', 'PluginTarget', 'RendererHandlers', 'plugin']
+__all__ = [
+    'DeclarativePluginSpec',
+    'InlineDelimited',
+    'RendererHandlers',
+    'RenderTemplate',
+    'install_declarative',
+    'plugin',
+]